requirements.txt
flask>=2.3.0
flask-cors>=3.0.10
pillow>=9.5.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.2
timm>=0.9.10
numpy>=1.24.0
werkzeug>=2.3.0
requests>=2.26.0
tqdm>=4.66.1
opencv-python>=4.8.1
scikit-image>=0.22.0
exifread>=3.0.0
psutil>=5.8.0
prometheus-client>=0.11.0
py-spy>=0.3.14
captum>=0.4.0
matplotlib>=3.4.3
shap>=0.43.0
albumentations>=1.3.1
pandas>=2.1.0
seaborn>=0.13.0
fpdf>=1.7.2
xlsxwriter>=3.1.9
plotly>=5.1.0
python-dotenv>=0.19.0
scikit-learn>=1.3.0
scipy>=1.11.0
schedule>=1.1.0
